Mudanças ao tipo usuário e post além como colocar os header files no seu devido sitio

+++ include/heap.h
@@ -1,7 +1,7 @@
 #ifndef HEAP_H
 #define HEAP_H
 
-#include <stdlib.h>
+//#include <stdlib.h>
 
 #define pai(i) (i-1) / 2
 #define esq(i) 2*i + 1
@@ -13,7 +13,7 @@ HEAP create_H(void);
 void destroy_H(HEAP x);
 int length_H(HEAP x);
 
-void add_Heap( HEAP x, int key , void* n , size_t fat);
-void* rem_Heap( HEAP x, int *key, size_t *fat );
+void add_Heap( HEAP x, int key , void* n , unsigned long fat);
+void* rem_Heap( HEAP x, int *key, unsigned long *fat );
 
 #endif
